exercise 4.21 completed

+++ osa4/blogilista/controllers/blogs.js
@@ -72,8 +72,19 @@ blogsRouter.put('/:id', async (request, response, next) => {
     })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    // await Blog.findByIdAndDelete(request.params.id)
+    // response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'only the owner can delete a blog ' })
+    }
 })
 
 module.exports = blogsRouter
